chore(error-tracking): make V3 the only error tracking query path
Remove the v1/v2 query builders, the useQueryV2/useQueryV3 schema fields, and
the error-tracking-query-v2/v3 + force-query feature flags. The denormalized
ClickHouse query (formerly 'v3') is now the single, unconditional path:

- backend: delete the v1/v2 builders, rename the v3 builder to
  ErrorTrackingQueryBuilder, and have the runner always use it
- api: the /error_tracking/query endpoints and the dashboards error-tracking
  list widget always resolve fingerprints and run V3 (no flag, no fallback)
- frontend: drop the query-version switch, flags, and wiring
- tests: consolidate the runner test suites into one and drop the flag-gated
  branches in the API tests
