+++ frontend/src/queries/schema.json
@@ -17304,7 +17304,7 @@
                     "type": "string"
                 },
                 "pendingFingerprintIssueStateUpdates": {
-                    "description": "Pending fingerprint issue state updates UNIONed into the fingerprint issue state subquery (V3 only). The backend caps the list at 50 entries; extras are dropped silently.",
+                    "description": "Pending fingerprint issue state updates UNIONed into the fingerprint issue state subquery. The backend caps the list at 50 entries; extras are dropped silently.",
                     "items": {
                         "$ref": "#/definitions/ErrorTrackingPendingFingerprintIssueStateUpdate"
                     },
@@ -17326,14 +17326,6 @@
                 },
                 "tags": {
                     "$ref": "#/definitions/QueryLogTags"
-                },
-                "useQueryV2": {
-                    "description": "Use V2 query path (ClickHouse postgres connector join instead of separate Postgres queries)",
-                    "type": "boolean"
-                },
-                "useQueryV3": {
-                    "description": "Use V3 query path (denormalized ClickHouse table, no Postgres joins)",
-                    "type": "boolean"
                 },
                 "version": {
                     "description": "version of the node, used for schema migrations",
