+++ products/error_tracking/backend/api/test/test_query_api.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from freezegun import freeze_time
 from posthog.test.base import APIBaseTest, ClickhouseTestMixin, _create_event, _create_person, flush_persons_and_events
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 from django.utils.timezone import now
 
@@ -21,6 +21,7 @@ from products.error_tracking.backend.models import (
     ErrorTrackingIssue,
     ErrorTrackingIssueAssignment,
     ErrorTrackingIssueFingerprintV2,
+    sync_issues_to_clickhouse,
 )
 
 
@@ -74,11 +75,6 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
 
     def setUp(self) -> None:
         super().setUp()
-        self.feature_flag_patcher = patch(
-            "products.error_tracking.backend.api.query.posthoganalytics.feature_enabled", return_value=False
-        )
-        self.mock_feature_enabled: Mock = self.feature_flag_patcher.start()
-        self.addCleanup(self.feature_flag_patcher.stop)
         _create_person(team=self.team, distinct_ids=["user-1"], is_identified=True)
         flush_persons_and_events()
 
@@ -87,6 +83,8 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
         ErrorTrackingIssueFingerprintV2.objects.create(
             team=self.team, issue=issue, fingerprint=fingerprint or self.fingerprint
         )
+        # the query always uses the denormalized ClickHouse table, so mirror the issue state there
+        sync_issues_to_clickhouse(issue_ids=[issue.id], team_id=self.team.pk)
         return issue
 
     def create_exception_event(
@@ -237,13 +235,10 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
         assert response.status_code == 200
         assert observed_tags == [(Product.ERROR_TRACKING, Feature.QUERY)]
 
-    def test_issues_list_uses_v3_when_force_flag_enabled(self) -> None:
-        self.mock_feature_enabled.return_value = True
-        observed_use_query_v3: list[bool | None] = []
+    def test_issues_list_normalizes_volume_resolution(self) -> None:
         observed_volume_resolutions: list[int] = []
 
         def calculate(runner: ErrorTrackingQueryRunner) -> FakeQueryResponse:
-            observed_use_query_v3.append(runner.query.useQueryV3)
             observed_volume_resolutions.append(runner.query.volumeResolution)
             return FakeQueryResponse({"results": [], "hasMore": False, "limit": 25, "offset": 0})
 
@@ -255,34 +250,7 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
             )
 
         assert response.status_code == 200
-        assert observed_use_query_v3 == [True]
         assert observed_volume_resolutions == [1]
-        feature_flag_call = self.mock_feature_enabled.call_args
-        assert feature_flag_call is not None
-        assert feature_flag_call.kwargs["groups"] == {
-            "organization": str(self.organization.id),
-            "project": str(self.team.id),
-        }
-
-    def test_issues_list_keeps_legacy_query_when_force_flag_disabled(self) -> None:
-        observed_use_query_v3: list[bool | None] = []
-        observed_volume_resolutions: list[int] = []
-
-        def calculate(runner: ErrorTrackingQueryRunner) -> FakeQueryResponse:
-            observed_use_query_v3.append(runner.query.useQueryV3)
-            observed_volume_resolutions.append(runner.query.volumeResolution)
-            return FakeQueryResponse({"results": [], "hasMore": False, "limit": 25, "offset": 0})
-
-        with patch("products.error_tracking.backend.api.query.ErrorTrackingQueryRunner.calculate", calculate):
-            response = self.client.post(
-                f"/api/environments/{self.team.id}/error_tracking/query/issues",
-                data={"volumeResolution": 0},
-                format="json",
-            )
-
-        assert response.status_code == 200
-        assert observed_use_query_v3 == [None]
-        assert observed_volume_resolutions == [0]
 
     def test_issue_detail_tags_clickhouse_queries(self) -> None:
         self.create_issue()
@@ -317,15 +285,12 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
         assert response.status_code == 200
         assert observed_tags == [(Product.ERROR_TRACKING, Feature.QUERY), (Product.ERROR_TRACKING, Feature.QUERY)]
 
-    def test_issue_detail_uses_v3_when_force_flag_enabled(self) -> None:
-        self.mock_feature_enabled.return_value = True
+    def test_issue_detail_filters_by_fingerprint(self) -> None:
         self.create_issue()
-        observed_use_query_v3: list[bool | None] = []
         observed_volume_resolutions: list[int] = []
         observed_filter_groups: list[dict[str, object] | None] = []
 
         def calculate_issue(runner: ErrorTrackingQueryRunner) -> FakeQueryResponse:
-            observed_use_query_v3.append(runner.query.useQueryV3)
             observed_volume_resolutions.append(runner.query.volumeResolution)
             observed_filter_groups.append(
                 runner.query.filterGroup.model_dump(mode="json") if runner.query.filterGroup else None
@@ -352,19 +317,19 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
             )
 
         assert response.status_code == 200
-        assert observed_use_query_v3 == [True]
         assert observed_volume_resolutions == [1]
         assert observed_filter_groups[0] is not None
         assert "$exception_fingerprint" in str(observed_filter_groups[0])
         assert self.fingerprint in str(observed_filter_groups[0])
 
-    def test_issue_detail_keeps_legacy_query_without_fingerprints(self) -> None:
-        self.mock_feature_enabled.return_value = True
+    def test_issue_detail_without_fingerprints_has_no_filter_group(self) -> None:
         ErrorTrackingIssue.objects.create(id=self.issue_id, team=self.team, name="TypeError")
-        observed_use_query_v3: list[bool | None] = []
+        observed_filter_groups: list[dict[str, object] | None] = []
 
         def calculate_issue(runner: ErrorTrackingQueryRunner) -> FakeQueryResponse:
-            observed_use_query_v3.append(runner.query.useQueryV3)
+            observed_filter_groups.append(
+                runner.query.filterGroup.model_dump(mode="json") if runner.query.filterGroup else None
+            )
             return FakeQueryResponse(
                 {
                     "results": [
@@ -387,7 +352,7 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
             )
 
         assert response.status_code == 200
-        assert observed_use_query_v3 == [None]
+        assert observed_filter_groups == [None]
 
     @freeze_time("2026-04-24T12:00:00Z")
     def test_issue_detail_returns_impact_top_frame_and_latest_release(self) -> None:
@@ -524,8 +489,7 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
         assert observed_tags == [(Product.ERROR_TRACKING, Feature.QUERY)]
 
     @freeze_time("2026-04-24T12:00:00Z")
-    def test_issue_events_uses_fingerprint_filter_when_force_flag_enabled(self) -> None:
-        self.mock_feature_enabled.return_value = True
+    def test_issue_events_matches_by_fingerprint(self) -> None:
         self.create_issue()
         self.create_exception_event(
             issue_id="01936e80-45e5-70bd-baa1-bf2f2ca4c532",
@@ -543,44 +507,9 @@ class TestErrorTrackingQueryAPI(ClickhouseTestMixin, APIBaseTest):
         assert response.json()["results"][0]["properties"]["$session_id"] == "session-id-1"
 
     @freeze_time("2026-04-24T12:00:00Z")
-    def test_issue_events_uses_fingerprint_filter_without_legacy_issue_id(self) -> None:
-        self.mock_feature_enabled.return_value = True
+    def test_issue_events_matches_events_without_issue_id(self) -> None:
         self.create_issue()
         self.create_exception_event(include_issue_id=False, properties={"$session_id": "session-id-1"})
-        flush_persons_and_events()
-
-        response = self.client.post(
-            f"/api/environments/{self.team.id}/error_tracking/query/issue_events",
-            data={"issueId": self.issue_id, "dateRange": {"date_from": "-1d", "date_to": "2026-04-25T00:00:00Z"}},
-            format="json",
-        )
-
-        assert response.status_code == 200
-        assert response.json()["results"][0]["properties"]["$session_id"] == "session-id-1"
-
-    @freeze_time("2026-04-24T12:00:00Z")
-    def test_issue_events_keeps_issue_id_filter_when_force_flag_disabled(self) -> None:
-        self.create_issue()
-        self.create_exception_event(
-            issue_id="01936e80-45e5-70bd-baa1-bf2f2ca4c532",
-            properties={"$session_id": "session-id-1"},
-        )
-        flush_persons_and_events()
-
-        response = self.client.post(
-            f"/api/environments/{self.team.id}/error_tracking/query/issue_events",
-            data={"issueId": self.issue_id, "dateRange": {"date_from": "-1d", "date_to": "2026-04-25T00:00:00Z"}},
-            format="json",
-        )
-
-        assert response.status_code == 200
-        assert response.json()["results"] == []
-
-    @freeze_time("2026-04-24T12:00:00Z")
-    def test_issue_events_uses_issue_id_filter_without_fingerprints(self) -> None:
-        self.mock_feature_enabled.return_value = True
-        ErrorTrackingIssue.objects.create(id=self.issue_id, team=self.team, name="TypeError")
-        self.create_exception_event(properties={"$session_id": "session-id-1"})
         flush_persons_and_events()
 
         response = self.client.post(
