+++ products/error_tracking/backend/facade/import_map.md
@@ -10,7 +10,7 @@ This tracks cross-product imports that currently bypass the facade.
 | `backend.models`                                                | `posthog/models/remote_config.py`, `posthog/models/product_intent/product_intent.py`, `posthog/management/commands/*`, plus tests/admin/demo                                                   | read/list/detail, counters, async/task |
 | `backend.hogql_queries.*`                                       | `posthog/hogql_queries/query_runner.py`, `products/signals/backend/temporal/backfill_error_tracking.py`, `products/posthog_ai/scripts/hogql_example/__init__.py`                               | read/list analytics query execution    |
 | `backend.api` package                                           | `posthog/api/__init__.py`                                                                                                                                                                      | HTTP routing entrypoint                |
-| `backend.api.issues` serializer                                 | `backend/hogql_queries/error_tracking_query_runner_v1.py`, `backend/hogql_queries/error_tracking_issue_correlation_query_runner.py`                                                            | read/list serialization coupling       |
+| `backend.api.issues` serializer                                 | `backend/hogql_queries/error_tracking_issue_correlation_query_runner.py`                                                                                                                       | read/list serialization coupling       |
 | `backend.remote_config` and `backend.api.suppression_rules`     | `posthog/models/remote_config.py`                                                                                                                                                              | config read                            |
 | `backend.embedding`, `backend.indexed_embedding`, `backend.sql` | `posthog/clickhouse/schema.py`, `posthog/clickhouse/migrations/*`, `posthog/api/embedding_worker.py`, `products/signals/backend/management/commands/cleanup_signals.py`, `posthog/conftest.py` | schema/constants (infra)               |
 | `backend.tools.search_issues`                                   | `ee/hogai/core/agent_modes/presets/error_tracking.py`                                                                                                                                          | tool invocation                        |
