+++ products/error_tracking/backend/hogql_queries/error_tracking_query_runner.py
@@ -15,10 +15,8 @@ from posthog.hogql_queries.query_runner import AnalyticsQueryRunner
 from posthog.models.filters.mixins.utils import cached_property
 from posthog.utils import relative_date_parse
 
+from products.error_tracking.backend.hogql_queries.error_tracking_query_builder import ErrorTrackingQueryBuilder
 from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_utils import validate_uuid_param
-from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_v1 import ErrorTrackingQueryV1Builder
-from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_v2 import ErrorTrackingQueryV2Builder
-from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_v3 import ErrorTrackingQueryV3Builder
 
 
 class ErrorTrackingQueryRunner(AnalyticsQueryRunner[ErrorTrackingQueryResponse]):
@@ -52,12 +50,8 @@ class ErrorTrackingQueryRunner(AnalyticsQueryRunner[ErrorTrackingQueryResponse])
             self.query.withLastEvent = False
 
     @cached_property
-    def _builder(self) -> ErrorTrackingQueryV1Builder | ErrorTrackingQueryV2Builder | ErrorTrackingQueryV3Builder:
-        if self.query.useQueryV3:
-            return ErrorTrackingQueryV3Builder(self.query, self.team, self.date_from, self.date_to)
-        if self.query.useQueryV2:
-            return ErrorTrackingQueryV2Builder(self.query, self.date_from, self.date_to)
-        return ErrorTrackingQueryV1Builder(self.query, self.team, self.date_from, self.date_to)
+    def _builder(self) -> ErrorTrackingQueryBuilder:
+        return ErrorTrackingQueryBuilder(self.query, self.team, self.date_from, self.date_to)
 
     def get_cache_payload(self) -> dict:
         payload = super().get_cache_payload()
