+++ products/error_tracking/backend/hogql_queries/test/test_error_tracking_query_runner.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
-from datetime import datetime, timedelta
+from datetime import UTC, datetime, timedelta
 from zoneinfo import ZoneInfo
 
 from freezegun import freeze_time
 from posthog.test.base import (
-    APIBaseTest,
     ClickhouseTestMixin,
+    NonAtomicBaseTestKeepIdentities,
     _create_event,
     _create_person,
     flush_persons_and_events,
@@ -32,9 +32,9 @@ from posthog.schema import (
     PropertyOperator,
 )
 
-from posthog.models import Organization, Team, User
 from posthog.models.utils import uuid7
 
+from products.error_tracking.backend.hogql_queries.error_tracking_query_builder import ErrorTrackingQueryBuilder
 from products.error_tracking.backend.hogql_queries.error_tracking_query_runner import ErrorTrackingQueryRunner
 from products.error_tracking.backend.hogql_queries.error_tracking_query_runner_utils import search_tokenizer
 from products.error_tracking.backend.models import (
@@ -42,27 +42,14 @@ from products.error_tracking.backend.models import (
     ErrorTrackingIssueAssignment,
     ErrorTrackingIssueFingerprintV2,
     override_error_tracking_issue_fingerprint,
+    sync_issues_to_clickhouse,
     update_error_tracking_issue_fingerprints,
 )
 
 from ee.models.rbac.role import Role
 
 
-class ErrorTrackingQueryRunnerTestsMixin:
-    __test__ = False
-
-    use_v2: bool = False
-    use_v3: bool = False
-
-    # will be provided by inheritance from PostHogTestCase
-    team: Team
-    user: User
-    organization: Organization
-
-    def assertEqual(
-        self, first: object, second: object, msg: object = None
-    ) -> None: ...  # will be provided by inheritance from TestCase
-
+class TestErrorTrackingQueryRunner(ClickhouseTestMixin, NonAtomicBaseTestKeepIdentities):
     distinct_id_one = "user_1"
     distinct_id_two = "user_2"
 
@@ -83,11 +70,13 @@ class ErrorTrackingQueryRunnerTestsMixin:
         override_error_tracking_issue_fingerprint(
             team_id=self.team.pk, fingerprint=fingerprint, issue_id=issue_id, version=version
         )
+        # reflect the new fingerprint -> issue mapping in the denormalized ClickHouse table
+        sync_issues_to_clickhouse(issue_ids=[issue_id], team_id=self.team.pk)
 
     def create_issue(self, issue_id, fingerprint, name=None, status=ErrorTrackingIssue.Status.ACTIVE):
         issue = ErrorTrackingIssue.objects.create(id=issue_id, team=self.team, status=status, name=name)
         ErrorTrackingIssueFingerprintV2.objects.create(team=self.team, issue=issue, fingerprint=fingerprint)
-
+        sync_issues_to_clickhouse(issue_ids=[issue.id], team_id=self.team.pk)
         return issue
 
     def create_events_and_issue(
@@ -121,6 +110,8 @@ class ErrorTrackingQueryRunnerTestsMixin:
                 person_id=person_id,
             )
 
+        sync_issues_to_clickhouse(issue_ids=[issue_id], team_id=self.team.pk)
+
     @classmethod
     def setUpClass(cls) -> None:
         from ee.clickhouse.materialized_columns.columns import get_materialized_columns, materialize
@@ -129,10 +120,10 @@ class ErrorTrackingQueryRunnerTestsMixin:
         for property_name in ("$exception_issue_id", "$exception_types", "$exception_values"):
             if (property_name, "properties") not in materialized_columns:
                 materialize("events", property_name, is_nullable=property_name == "$exception_issue_id")
-        super(ErrorTrackingQueryRunnerTestsMixin, cls).setUpClass()  # type: ignore[misc] # noqa: UP008
+        super().setUpClass()
 
     def setUp(self):
-        super(ErrorTrackingQueryRunnerTestsMixin, self).setUp()  # type: ignore[misc] # noqa: UP008
+        super().setUp()
 
         with freeze_time("2020-01-10 12:11:00"):
             _create_person(
@@ -209,71 +200,76 @@ class ErrorTrackingQueryRunnerTestsMixin:
                     personId=personId,
                     groupKey=groupKey,
                     groupTypeIndex=groupTypeIndex,
-                    useQueryV2=self.use_v2,
-                    useQueryV3=self.use_v3,
                 ),
             )
             .calculate()
             .model_dump()
         )
 
+    @parameterized.expand(
+        [
+            (
+                "default",
+                {},
+                [
+                    "id",
+                    "status",
+                    "name",
+                    "description",
+                    "assignee_user_id",
+                    "assignee_role_id",
+                    "last_seen",
+                    "first_seen",
+                    "function",
+                    "source",
+                    "library",
+                ],
+            ),
+            (
+                "with_aggregations",
+                {"withAggregations": True},
+                [
+                    "id",
+                    "status",
+                    "name",
+                    "description",
+                    "assignee_user_id",
+                    "assignee_role_id",
+                    "last_seen",
+                    "first_seen",
+                    "function",
+                    "source",
+                    "occurrences",
+                    "sessions",
+                    "users",
+                    "volumeRange",
+                    "library",
+                ],
+            ),
+            (
+                "with_first_event",
+                {"withFirstEvent": True},
+                [
+                    "id",
+                    "status",
+                    "name",
+                    "description",
+                    "assignee_user_id",
+                    "assignee_role_id",
+                    "last_seen",
+                    "first_seen",
+                    "function",
+                    "source",
+                    "first_event",
+                    "library",
+                ],
+            ),
+        ]
+    )
     @freeze_time("2022-01-10T12:11:00")
-    @snapshot_clickhouse_queries
-    def test_column_names(self):
-        columns = self._calculate()["columns"]
-        self.assertEqual(
-            columns,
-            ["id", "last_seen", "first_seen", "function", "source", "library"],
-        )
-
-        columns = self._calculate(withAggregations=True)["columns"]
-        self.assertEqual(
-            columns,
-            [
-                "id",
-                "last_seen",
-                "first_seen",
-                "function",
-                "source",
-                "occurrences",
-                "sessions",
-                "users",
-                "volumeRange",
-                "library",
-            ],
-        )
-
-        columns = self._calculate(withFirstEvent=True)["columns"]
-        self.assertEqual(
-            columns,
-            [
-                "id",
-                "last_seen",
-                "first_seen",
-                "function",
-                "source",
-                "first_event",
-                "library",
-            ],
-        )
-
-        columns = self._calculate(issueId=self.issue_id_one, withAggregations=True, withFirstEvent=True)["columns"]
-        self.assertEqual(
-            columns,
-            [
-                "id",
-                "last_seen",
-                "first_seen",
-                "function",
-                "source",
-                "occurrences",
-                "sessions",
-                "users",
-                "volumeRange",
-                "first_event",
-                "library",
-            ],
-        )
+    def test_column_names(self, _name, kwargs, expected_columns):
+        columns = self._calculate(**kwargs)["columns"]
+        self.assertEqual(columns, expected_columns)
 
     @freeze_time("2022-01-10T12:11:00")
     def test_date_range_resolution(self):
@@ -420,24 +416,28 @@ class ErrorTrackingQueryRunnerTestsMixin:
     @freeze_time("2022-01-10 12:11:00")
     @snapshot_clickhouse_queries
     def test_correctly_counts_session_ids(self):
+        common_properties = {
+            "$exception_issue_id": self.issue_id_one,
+            "$exception_fingerprint": self.issue_one_fingerprint,
+        }
         _create_event(
             distinct_id=self.distinct_id_one,
             event="$exception",
             team=self.team,
-            properties={"$session_id": str(uuid7()), "$exception_issue_id": self.issue_id_one},
+            properties={"$session_id": str(uuid7()), **common_properties},
         )
         _create_event(
             distinct_id=self.distinct_id_one,
             event="$exception",
             team=self.team,
-            properties={"$session_id": str(uuid7()), "$exception_issue_id": self.issue_id_one},
+            properties={"$session_id": str(uuid7()), **common_properties},
         )
         # blank string
         _create_event(
             distinct_id=self.distinct_id_one,
             event="$exception",
             team=self.team,
-            properties={"$session_id": "", "$exception_issue_id": self.issue_id_one},
+            properties={"$session_id": "", **common_properties},
         )
         flush_persons_and_events()
 
@@ -484,11 +484,12 @@ class ErrorTrackingQueryRunnerTestsMixin:
         self.assertEqual([r["id"] for r in results], [self.issue_id_one, self.issue_id_two, self.issue_id_three])
 
     @freeze_time("2022-01-10T12:11:00")
-    @snapshot_clickhouse_queries
     def test_status(self):
         resolved_issue = ErrorTrackingIssue.objects.get(id=self.issue_id_one)
         resolved_issue.status = ErrorTrackingIssue.Status.RESOLVED
         resolved_issue.save()
+        # re-sync after status change
+        sync_issues_to_clickhouse(issue_ids=[self.issue_id_one], team_id=self.team.pk)
 
         results = self._calculate(status="active")["results"]
         self.assertEqual([r["id"] for r in results], [self.issue_id_three, self.issue_id_two])
@@ -517,7 +518,6 @@ class ErrorTrackingQueryRunnerTestsMixin:
         self.assertEqual(results[1]["aggregations"]["occurrences"], 1)
 
     @freeze_time("2022-01-10T12:11:00")
-    @snapshot_clickhouse_queries
     def test_user_assignee(self):
         issue_id = "e9ac529f-ac1c-4a96-bd3a-107034368d64"
         self.create_events_and_issue(
@@ -527,12 +527,14 @@ class ErrorTrackingQueryRunnerTestsMixin:
         )
         flush_persons_and_events()
         ErrorTrackingIssueAssignment.objects.create(issue_id=issue_id, user=self.user, team=self.team)
+        # re-sync with a newer version so the assignment wins argMax over the create-time row
+        with freeze_time("2022-01-10T12:11:05"):
+            sync_issues_to_clickhouse(issue_ids=[issue_id], team_id=self.team.pk)
 
         results = self._calculate(assignee={"type": "user", "id": self.user.pk})["results"]
         self.assertEqual([x["id"] for x in results], [issue_id])
 
     @freeze_time("2022-01-10T12:11:00")
-    @snapshot_clickhouse_queries
     def test_role_assignee(self):
         issue_id = "e9ac529f-ac1c-4a96-bd3a-107034368d64"
         self.create_events_and_issue(
@@ -543,6 +545,9 @@ class ErrorTrackingQueryRunnerTestsMixin:
         flush_persons_and_events()
         role = Role.objects.create(name="Test Team", organization=self.organization)
         ErrorTrackingIssueAssignment.objects.create(issue_id=issue_id, role=role, team=self.team)
+        # re-sync with a newer version so the assignment wins argMax over the create-time row
+        with freeze_time("2022-01-10T12:11:05"):
+            sync_issues_to_clickhouse(issue_ids=[issue_id], team_id=self.team.pk)
 
         results = self._calculate(assignee={"type": "role", "id": str(role.id)})["results"]
         self.assertEqual([x["id"] for x in results], [issue_id])
@@ -566,6 +571,145 @@ class ErrorTrackingQueryRunnerTestsMixin:
             )
         )["results"]
         self.assertEqual(len(results), 1)
+
+    @parameterized.expand(
+        [
+            (
+                "or_returns_union",
+                FilterLogicalOperator.OR_,
+                # issue_one (TypeError) and issue_two (ReferenceError) both match
+                [True, True, False],
+            ),
+            (
+                "and_returns_intersection",
+                FilterLogicalOperator.AND_,
+                # No issue has both names — AND yields empty
+                [False, False, False],
+            ),
+        ]
+    )
+    @freeze_time("2022-01-10T12:11:00")
+    def test_filter_group_operator(self, _name, operator: FilterLogicalOperator, expected_membership: list[bool]):
+        results = self._calculate(
+            filterGroup=PropertyGroupFilter(
+                type=FilterLogicalOperator.AND_,
+                values=[
+                    PropertyGroupFilterValue(
+                        type=operator,
+                        values=[
+                            ErrorTrackingIssueFilter(
+                                key="name", value=[self.issue_name_one], operator=PropertyOperator.EXACT
+                            ),
+                            ErrorTrackingIssueFilter(
+                                key="name", value=[self.issue_name_two], operator=PropertyOperator.EXACT
+                            ),
+                        ],
+                    )
+                ],
+            )
+        )["results"]
+        result_ids = {r["id"] for r in results}
+        expected_ids = {
+            issue_id
+            for issue_id, included in zip(
+                [self.issue_id_one, self.issue_id_two, self.issue_id_three], expected_membership
+            )
+            if included
+        }
+        self.assertEqual(result_ids, expected_ids)
+
+    @freeze_time("2022-01-10T12:11:00")
+    def test_nested_filter_group_routes_issue_filters_to_issue_fields(self):
+        filter_group = PropertyGroupFilter(
+            type=FilterLogicalOperator.AND_,
+            values=[
+                PropertyGroupFilterValue(
+                    type=FilterLogicalOperator.AND_,
+                    values=[
+                        PropertyGroupFilterValue(
+                            type=FilterLogicalOperator.OR_,
+                            values=[
+                                EventPropertyFilter(key="$browser", value=["Firefox"], operator=PropertyOperator.EXACT),
+                                EventPropertyFilter(key="$browser", value=["Chrome"], operator=PropertyOperator.EXACT),
+                                ErrorTrackingIssueFilter(
+                                    key="name", value=[self.issue_name_one], operator=PropertyOperator.EXACT
+                                ),
+                            ],
+                        ),
+                        EventPropertyFilter(
+                            key="$exception_issue_id", value=[self.issue_id_one], operator=PropertyOperator.EXACT
+                        ),
+                    ],
+                )
+            ],
+        )
+
+        builder = ErrorTrackingQueryBuilder(
+            query=ErrorTrackingQuery(
+                kind="ErrorTrackingQuery",
+                dateRange=DateRange(date_from="-7d"),
+                filterGroup=filter_group,
+                orderBy="last_seen",
+                volumeResolution=1,
+            ),
+            team=self.team,
+            date_from=datetime(2022, 1, 3, tzinfo=UTC),
+            date_to=datetime(2022, 1, 10, tzinfo=UTC),
+        )
+        user_filter_expr = builder._user_filter_expr()
+        assert user_filter_expr is not None
+        user_filter_hogql = user_filter_expr.to_hogql()
+        self.assertIn("e.issue_name", user_filter_hogql)
+        self.assertNotIn("properties.name", user_filter_hogql)
+
+        results = self._calculate(filterGroup=filter_group)["results"]
+        self.assertEqual([r["id"] for r in results], [self.issue_id_one])
+
+    @freeze_time("2022-01-10T12:11:00")
+    def test_event_filter_group_operator(self):
+        firefox_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5333"
+        chrome_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5334"
+        safari_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5335"
+        self.create_events_and_issue(
+            issue_id=firefox_issue_id,
+            fingerprint="firefox_issue_fingerprint",
+            distinct_ids=[self.distinct_id_one],
+            additional_properties={"$browser": "Firefox"},
+        )
+        self.create_events_and_issue(
+            issue_id=chrome_issue_id,
+            fingerprint="chrome_issue_fingerprint",
+            distinct_ids=[self.distinct_id_one],
+            additional_properties={"$browser": "Chrome"},
+        )
+        self.create_events_and_issue(
+            issue_id=safari_issue_id,
+            fingerprint="safari_issue_fingerprint",
+            distinct_ids=[self.distinct_id_one],
+            additional_properties={"$browser": "Safari"},
+        )
+        flush_persons_and_events()
+
+        browser_filters = [
+            EventPropertyFilter(key="$browser", value=["Firefox"], operator=PropertyOperator.EXACT),
+            EventPropertyFilter(key="$browser", value=["Chrome"], operator=PropertyOperator.EXACT),
+        ]
+
+        or_results = self._calculate(
+            filterGroup=PropertyGroupFilter(
+                type=FilterLogicalOperator.AND_,
+                values=[PropertyGroupFilterValue(type=FilterLogicalOperator.OR_, values=browser_filters)],
+            )
+        )["results"]
+        self.assertEqual({result["id"] for result in or_results}, {firefox_issue_id, chrome_issue_id})
+
+        and_results = self._calculate(
+            filterGroup=PropertyGroupFilter(
+                type=FilterLogicalOperator.AND_,
+                values=[PropertyGroupFilterValue(type=FilterLogicalOperator.AND_, values=browser_filters)],
+            )
+        )["results"]
+        self.assertEqual([result["id"] for result in and_results], [])
 
     @freeze_time("2022-01-10T12:11:00")
     @snapshot_clickhouse_queries
@@ -739,11 +883,6 @@ class ErrorTrackingQueryRunnerTestsMixin:
         self.assertEqual(sum(first_aggregations["volumeRange"]), 24 * 5)
         self.assertEqual(first_aggregations["volumeRange"], [60, 60, 0, 0])
 
-
-class TestErrorTrackingQueryRunner(ErrorTrackingQueryRunnerTestsMixin, ClickhouseTestMixin, APIBaseTest):
-    __test__ = True
-    use_v2 = False
-
     @parameterized.expand(["issueId", "personId"])
     def test_rejects_malformed_uuid_params(self, field):
         with self.assertRaises(ValidationError):
@@ -761,52 +900,6 @@ class TestErrorTrackingQueryRunner(ErrorTrackingQueryRunnerTestsMixin, Clickhous
             ),
         )
         self.assertEqual(runner.query.issueId, "01936e7f-d7ff-7314-b2d4-7627981e34f0")
-
-    @freeze_time("2022-01-10T12:11:00")
-    def test_event_filter_group_operator(self):
-        firefox_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5333"
-        chrome_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5334"
-        safari_issue_id = "01936e80-aa51-746f-aec4-cdf16a5c5335"
-        self.create_events_and_issue(
-            issue_id=firefox_issue_id,
-            fingerprint="firefox_issue_fingerprint",
-            distinct_ids=[self.distinct_id_one],
-            additional_properties={"$browser": "Firefox"},
-        )
-        self.create_events_and_issue(
-            issue_id=chrome_issue_id,
-            fingerprint="chrome_issue_fingerprint",
-            distinct_ids=[self.distinct_id_one],
-            additional_properties={"$browser": "Chrome"},
-        )
-        self.create_events_and_issue(
-            issue_id=safari_issue_id,
-            fingerprint="safari_issue_fingerprint",
-            distinct_ids=[self.distinct_id_one],
-            additional_properties={"$browser": "Safari"},
-        )
-        flush_persons_and_events()
-
-        browser_filters = [
-            EventPropertyFilter(key="$browser", value=["Firefox"], operator=PropertyOperator.EXACT),
-            EventPropertyFilter(key="$browser", value=["Chrome"], operator=PropertyOperator.EXACT),
-        ]
-
-        or_results = self._calculate(
-            filterGroup=PropertyGroupFilter(
-                type=FilterLogicalOperator.AND_,
-                values=[PropertyGroupFilterValue(type=FilterLogicalOperator.OR_, values=browser_filters)],
-            )
-        )["results"]
-        self.assertEqual({result["id"] for result in or_results}, {firefox_issue_id, chrome_issue_id})
-
-        and_results = self._calculate(
-            filterGroup=PropertyGroupFilter(
-                type=FilterLogicalOperator.AND_,
-                values=[PropertyGroupFilterValue(type=FilterLogicalOperator.AND_, values=browser_filters)],
-            )
-        )["results"]
-        self.assertEqual([result["id"] for result in and_results], [])
 
 
 class TestSearchTokenizer(TestCase):
