+++ products/error_tracking/frontend/components/IssueQueryOptions/IssueQueryOptions.tsx
@@ -1,28 +1,19 @@
 import { useActions, useValues } from 'kea'
 
 import { IconRefresh } from '@posthog/icons'
-import { LemonButton, LemonMenu, LemonSelect, LemonSwitch, Spinner } from '@posthog/lemon-ui'
+import { LemonButton, LemonMenu, LemonSelect, Spinner } from '@posthog/lemon-ui'
 
 import { issuesDataNodeLogic } from '../../logics/issuesDataNodeLogic'
 import { ORDER_BY_OPTIONS, issueQueryOptionsLogic } from './issueQueryOptionsLogic'
 
 export const IssueQueryOptions = (): JSX.Element => {
-    const { orderBy, orderDirection, useQueryV3, showQueryV3Switch } = useValues(issueQueryOptionsLogic)
-    const { setOrderBy, setOrderDirection, setUseQueryV3 } = useActions(issueQueryOptionsLogic)
+    const { orderBy, orderDirection } = useValues(issueQueryOptionsLogic)
+    const { setOrderBy, setOrderDirection } = useActions(issueQueryOptionsLogic)
 
     return (
         <span className="flex items-center justify-between gap-2 self-end">
             <Reload />
             <div className="flex items-center gap-2 self-end">
-                {showQueryV3Switch && (
-                    <LemonSwitch
-                        checked={useQueryV3}
-                        onChange={setUseQueryV3}
-                        label="v3 query"
-                        size="small"
-                        tooltip="Use denormalized ClickHouse table (no Postgres joins)"
-                    />
-                )}
                 <div className="flex items-center gap-1">
                     <span>Sort by:</span>
