+++ products/error_tracking/frontend/components/IssueQueryOptions/issueQueryOptionsLogic.ts
@@ -1,10 +1,8 @@
 import equal from 'fast-deep-equal'
-import { actions, connect, kea, key, listeners, path, props, reducers, selectors } from 'kea'
+import { actions, kea, key, listeners, path, props, reducers } from 'kea'
 import { actionToUrl, router, urlToAction } from 'kea-router'
 import posthog from 'posthog-js'
 
-import { FEATURE_FLAGS } from 'lib/constants'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { Params } from 'scenes/sceneTypes'
 
 import { ErrorTrackingIssue, ErrorTrackingQuery } from '~/queries/schema/schema-general'
@@ -38,16 +36,11 @@ export const issueQueryOptionsLogic = kea<issueQueryOptionsLogicType>([
     props({} as IssueQueryOptionsLogicProps),
     key(({ logicKey }) => logicKey),
 
-    connect({
-        values: [featureFlagLogic, ['featureFlags']],
-    }),
-
     actions({
         setOrderBy: (orderBy: ErrorTrackingQueryOrderBy) => ({ orderBy }),
         setOrderDirection: (orderDirection: ErrorTrackingQueryOrderDirection) => ({ orderDirection }),
         setAssignee: (assignee: ErrorTrackingIssue['assignee']) => ({ assignee }),
         setStatus: (status: ErrorTrackingQueryStatus) => ({ status }),
-        setUseQueryV3: (useQueryV3: boolean) => ({ useQueryV3 }),
     }),
 
     reducers({
@@ -78,24 +71,6 @@ export const issueQueryOptionsLogic = kea<issueQueryOptionsLogicType>([
             {
                 setStatus: (_, { status }) => status,
             },
-        ],
-        useQueryV3: [
-            false as boolean,
-            { persist: true },
-            {
-                setUseQueryV3: (_, { useQueryV3 }) => useQueryV3,
-            },
-        ],
-    }),
-
-    selectors({
-        showQueryV3Switch: [
-            (s) => [s.featureFlags],
-            (featureFlags): boolean => !!featureFlags[FEATURE_FLAGS.ERROR_TRACKING_QUERY_V3],
-        ],
-        forceQueryV3: [
-            (s) => [s.featureFlags],
-            (featureFlags): boolean => !!featureFlags[FEATURE_FLAGS.ERROR_TRACKING_FORCE_QUERY_V3],
         ],
     }),
